# feat: baseten performance client in typescript (with rust bindings)