--- a/baseten-performance-client-typescript/baseten-performance-client/.cargo/config.toml
+++ b/baseten-performance-client-typescript/baseten-performance-client/.cargo/config.toml
@@ -1,0 +1,5 @@
+[target.aarch64-unknown-linux-musl]
+linker = "aarch64-linux-musl-gcc"
+rustflags = ["-C", "target-feature=-crt-static"]
+[target.x86_64-pc-windows-msvc]
+rustflags = ["-C", "target-feature=+crt-static"]