--- a/baseten-performance-client-typescript/baseten-performance-client/.npmignore
+++ b/baseten-performance-client-typescript/baseten-performance-client/.npmignore
@@ -1,0 +1,13 @@
+target
+Cargo.lock
+.cargo
+.github
+npm
+.eslintrc
+.prettierignore
+rustfmt.toml
+yarn.lock
+*.node
+.yarn
+__test__
+renovate.json