--- a/baseten-performance-client-typescript/baseten-performance-client/.yarnrc.yml
+++ b/baseten-performance-client-typescript/baseten-performance-client/.yarnrc.yml
@@ -1,0 +1,1 @@
+nodeLinker: node-modules