--- a/baseten-performance-client-typescript/baseten-performance-client/Cargo.toml
+++ b/baseten-performance-client-typescript/baseten-performance-client/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+edition = "2021"
+name = "truss_baseten-performance-client"
+version = "0.0.0"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+napi = { version = "2.12.2", features = ["napi8", "async", "serde-json"] }
+napi-derive = "2.12.2"
+futures = "0.3"
+tokio = { version = "1", features = ["full"] }
+tokio-util = { version = "0.7" }
+once_cell = "1.19"
+reqwest = { version = "0.11", features = ["json", "stream", "gzip", "brotli"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+url = "2.5"
+percent-encoding = "2.3"
+
+[build-dependencies]
+napi-build = "2.0.1"
+
+[profile.release]
+lto = true
+strip = "symbols"