--- a/baseten-performance-client-typescript/baseten-performance-client/Cargo.toml
+++ b/baseten-performance-client-typescript/baseten-performance-client/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
-name = "truss_baseten-performance-client"
-version = "0.0.0"
+name = "baseten-performance-client"
+version = "0.0.1"
 
 [lib]
 crate-type = ["cdylib"]