--- a/baseten-performance-client-typescript/baseten-performance-client/build.rs
+++ b/baseten-performance-client-typescript/baseten-performance-client/build.rs
@@ -1,0 +1,5 @@
+extern crate napi_build;
+
+fn main() {
+  napi_build::setup();
+}