--- a/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm-eabi/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm-eabi/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-android-arm-eabi`
+
+This is the **armv7-linux-androideabi** binary for `@truss/baseten-performance-client`