--- a/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm-eabi/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm-eabi/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-android-arm-eabi",
+  "version": "0.0.0",
+  "os": [
+    "android"
+  ],
+  "cpu": [
+    "arm"
+  ],
+  "main": "baseten-performance-client.android-arm-eabi.node",
+  "files": [
+    "baseten-performance-client.android-arm-eabi.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}