--- a/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm64/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm64/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-android-arm64`
+
+This is the **aarch64-linux-android** binary for `@truss/baseten-performance-client`