--- a/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm64/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/android-arm64/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-android-arm64",
+  "version": "0.0.0",
+  "os": [
+    "android"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "main": "baseten-performance-client.android-arm64.node",
+  "files": [
+    "baseten-performance-client.android-arm64.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}