--- a/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-arm64/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-arm64/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-darwin-arm64`
+
+This is the **aarch64-apple-darwin** binary for `@truss/baseten-performance-client`