--- a/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-universal/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-universal/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-darwin-universal`
+
+This is the **universal-apple-darwin** binary for `@truss/baseten-performance-client`