--- a/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-universal/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-universal/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@truss/baseten-performance-client-darwin-universal",
+  "version": "0.0.0",
+  "os": [
+    "darwin"
+  ],
+  "main": "baseten-performance-client.darwin-universal.node",
+  "files": [
+    "baseten-performance-client.darwin-universal.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}