--- a/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-x64/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-x64/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-darwin-x64`
+
+This is the **x86_64-apple-darwin** binary for `@truss/baseten-performance-client`