--- a/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-x64/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/darwin-x64/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-darwin-x64",
+  "version": "0.0.0",
+  "os": [
+    "darwin"
+  ],
+  "cpu": [
+    "x64"
+  ],
+  "main": "baseten-performance-client.darwin-x64.node",
+  "files": [
+    "baseten-performance-client.darwin-x64.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}