--- a/baseten-performance-client-typescript/baseten-performance-client/npm/freebsd-x64/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/freebsd-x64/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-freebsd-x64`
+
+This is the **x86_64-unknown-freebsd** binary for `@truss/baseten-performance-client`