--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-gnueabihf/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-gnueabihf/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-linux-arm-gnueabihf`
+
+This is the **armv7-unknown-linux-gnueabihf** binary for `@truss/baseten-performance-client`