--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-gnueabihf/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-gnueabihf/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-linux-arm-gnueabihf",
+  "version": "0.0.0",
+  "os": [
+    "linux"
+  ],
+  "cpu": [
+    "arm"
+  ],
+  "main": "baseten-performance-client.linux-arm-gnueabihf.node",
+  "files": [
+    "baseten-performance-client.linux-arm-gnueabihf.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}