--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-musleabihf/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-musleabihf/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-linux-arm-musleabihf`
+
+This is the **armv7-unknown-linux-musleabihf** binary for `@truss/baseten-performance-client`