--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-musleabihf/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm-musleabihf/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-linux-arm-musleabihf",
+  "version": "0.0.0",
+  "os": [
+    "linux"
+  ],
+  "cpu": [
+    "arm"
+  ],
+  "main": "baseten-performance-client.linux-arm-musleabihf.node",
+  "files": [
+    "baseten-performance-client.linux-arm-musleabihf.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}