--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm64-musl/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-arm64-musl/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@truss/baseten-performance-client-linux-arm64-musl",
+  "version": "0.0.0",
+  "os": [
+    "linux"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "main": "baseten-performance-client.linux-arm64-musl.node",
+  "files": [
+    "baseten-performance-client.linux-arm64-musl.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  },
+  "libc": [
+    "musl"
+  ]
+}