--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-riscv64-gnu/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-riscv64-gnu/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-linux-riscv64-gnu`
+
+This is the **riscv64gc-unknown-linux-gnu** binary for `@truss/baseten-performance-client`