--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-x64-gnu/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-x64-gnu/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-linux-x64-gnu`
+
+This is the **x86_64-unknown-linux-gnu** binary for `@truss/baseten-performance-client`