--- a/baseten-performance-client-typescript/baseten-performance-client/npm/linux-x64-musl/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/linux-x64-musl/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@truss/baseten-performance-client-linux-x64-musl",
+  "version": "0.0.0",
+  "os": [
+    "linux"
+  ],
+  "cpu": [
+    "x64"
+  ],
+  "main": "baseten-performance-client.linux-x64-musl.node",
+  "files": [
+    "baseten-performance-client.linux-x64-musl.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  },
+  "libc": [
+    "musl"
+  ]
+}