--- a/baseten-performance-client-typescript/baseten-performance-client/npm/win32-arm64-msvc/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/win32-arm64-msvc/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-win32-arm64-msvc",
+  "version": "0.0.0",
+  "os": [
+    "win32"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "main": "baseten-performance-client.win32-arm64-msvc.node",
+  "files": [
+    "baseten-performance-client.win32-arm64-msvc.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}