--- a/baseten-performance-client-typescript/baseten-performance-client/npm/win32-ia32-msvc/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/win32-ia32-msvc/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@truss/baseten-performance-client-win32-ia32-msvc",
+  "version": "0.0.0",
+  "os": [
+    "win32"
+  ],
+  "cpu": [
+    "ia32"
+  ],
+  "main": "baseten-performance-client.win32-ia32-msvc.node",
+  "files": [
+    "baseten-performance-client.win32-ia32-msvc.node"
+  ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 10"
+  }
+}