--- a/baseten-performance-client-typescript/baseten-performance-client/npm/win32-x64-msvc/README.md
+++ b/baseten-performance-client-typescript/baseten-performance-client/npm/win32-x64-msvc/README.md
@@ -1,0 +1,3 @@
+# `@truss/baseten-performance-client-win32-x64-msvc`
+
+This is the **x86_64-pc-windows-msvc** binary for `@truss/baseten-performance-client`