--- a/baseten-performance-client-typescript/baseten-performance-client/package.json
+++ b/baseten-performance-client-typescript/baseten-performance-client/package.json
@@ -30,7 +30,7 @@
     "typescript": "^5.8.3"
   },
   "engines": {
-    "node": "^22.14.0"
+    "node": ">= 10"
   },
   "scripts": {
     "artifacts": "napi artifacts",