--- a/baseten-performance-client-typescript/baseten-performance-client/rustfmt.toml
+++ b/baseten-performance-client-typescript/baseten-performance-client/rustfmt.toml
@@ -1,0 +1,2 @@
+tab_spaces = 2
+edition = "2021"