<!--
  What does this PR add, remove, and/or change?
-->
## :rocket: What

rust bindings for baseten-performance-client in typescript.
the only files that need to be reviewed are `src/lib.rs` and `__test__/test-baseten-performance-client.ts`. literally every other file was autogenerated by napi.

<!--
  How was the change described above implemented?
-->
## :computer: How

via napi-rs

<!--
  How have I ensured release and ongoing quality of this change?
-->
## :microscope: Testing

if running for the first time, run `yarn build`.

then, to run the test, run `npx ts-node __test__/test-baseten-performance-client.ts <BASE_URL_TO_YOUR_BASETEN_ENDPOINT>`.

<details>
<summary>sample output</summary>

```
🧪 Baseten Performance Client Integration Test

══════════════════════════════════════════════════════════════════════
Test 1: Basic Rerank Functionality
══════════════════════════════════════════════════════════════════════

✅ Basic rerank test passed
   Results: 5 documents ranked
   Top 3 results:
     1. Document 0: score = 0.9976
     2. Document 2: score = 0.0002
     3. Document 3: score = 0.0001

══════════════════════════════════════════════════════════════════════
Test 2: Performance Measurement
══════════════════════════════════════════════════════════════════════

📊 Run 1/3:
    Duration: 301.54ms
    Top result: Document 0 (score: 0.9976)
    Average batch time: 273.80ms

📊 Run 2/3:
    Duration: 288.44ms
    Top result: Document 0 (score: 0.9976)
    Average batch time: 259.26ms

📊 Run 3/3:
    Duration: 322.70ms
    Top result: Document 0 (score: 0.9976)
    Average batch time: 312.85ms

📈 Performance Summary:
   Average time per run: 304.23ms
   Documents processed: 15
   Average throughput: 49.3 docs/sec

   Note: PerformanceClient uses batching (4 docs/batch) with concurrent processing

══════════════════════════════════════════════════════════════════════
Test 3: Error Handling and Validation
══════════════════════════════════════════════════════════════════════

✅ Invalid API key: Correctly failed with: HTTP 403: API request failed with status 403 Forbidden: {
	"error": "please check the bearer token you provided"
}

✅ Empty documents array: Correctly failed with: Inputs list cannot be empty
✅ Empty document in array: Correctly failed with: Input string at index 0 cannot be empty

══════════════════════════════════════════════════════════════════════
Test 4: Performance at Different Scales
══════════════════════════════════════════════════════════════════════

📊 Testing with 25 documents:
    Duration: 495.12ms
    Throughput: 50.5 docs/sec
    Batches: 7 (7 expected)

📊 Testing with 50 documents:
    Duration: 419.37ms
    Throughput: 119.2 docs/sec
    Batches: 13 (13 expected)

📊 Testing with 100 documents:
    Duration: 434.23ms
    Throughput: 230.3 docs/sec
    Batches: 25 (25 expected)

📊 Testing with 200 documents:
    Duration: 582.42ms
    Throughput: 343.4 docs/sec
    Batches: 50 (50 expected)

══════════════════════════════════════════════════════════════════════
Test 5: Large Documents (100 docs, ~500 words each)
══════════════════════════════════════════════════════════════════════

📊 Testing with 100 long documents:
    Duration: 630.59ms
    Throughput: 158.6 docs/sec
    Batches: 25 (25 expected)
    Average batch time: 469.62ms


══════════════════════════════════════════════════════════════════════
✅ All tests completed successfully!
══════════════════════════════════════════════════════════════════════
```

</details>