@michaelfeil 2025-07-16
jfyi, @venkatesh-narayan adding a more modular structure here: https://github.com/basetenlabs/truss/pull/1779, which should enable better ts bindings with less code / performance regressions. 

Go you are about the gzip / brotoli features?